Change HALF_SECOND_MILLIS from a Long to an Int to match the other constants

+++ app/src/main/java/com/example/bronnbakestimer/logic/Constants.kt
@@ -60,7 +60,7 @@ object Constants {
      *
      * Value: 500 milliseconds (0.5 seconds)
      */
-    const val HALF_SECOND_MILLIS: Long = 500
+    const val HALF_SECOND_MILLIS: Int = 500
 
     /**
      * Number of updates per second for the timer's internal data.
